requirements.txt
PySide6~=6.10.0
psycopg2-binary